EPMLABSBRN-981 BE: add column withPictures for subgroup
- Addressed code review comments
