+++ src/main/kotlin/com/epam/brn/controller/SubGroupController.kt
@@ -2,16 +2,13 @@ package com.epam.brn.controller
 
 import com.epam.brn.dto.BaseResponseDto
 import com.epam.brn.dto.BaseSingleObjectResponseDto
-import com.epam.brn.dto.request.SubGroupChangeRequest
 import com.epam.brn.service.SubGroupService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
@@ -40,15 +37,5 @@ class SubGroupController(private val subGroupsService: SubGroupService) {
     fun deleteSubGroupById(@PathVariable(value = "subGroupId") subGroupId: Long): ResponseEntity<BaseSingleObjectResponseDto> {
         subGroupsService.deleteSubGroupById(subGroupId)
         return ResponseEntity.ok(BaseSingleObjectResponseDto(data = Unit))
-    }
-
-    @PatchMapping("{subGroupId}")
-    @ApiOperation("Update subgroup by id")
-    fun updateSubGroupById(
-        @PathVariable(value = "subGroupId") subGroupId: Long,
-        @RequestBody subGroup: SubGroupChangeRequest
-    ): ResponseEntity<BaseSingleObjectResponseDto> {
-        val subGroupDto = subGroupsService.updateSubGroupById(subGroupId, subGroup)
-        return ResponseEntity.ok(BaseSingleObjectResponseDto(data = subGroupDto))
     }
 }
